+++ test/core/store/test_store.py
@@ -2,6 +2,7 @@
 # Permissions are hereby granted under the terms of the MIT License:
 # https://opensource.org/licenses/MIT.
 import unittest
+from typing import Literal
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -9,10 +10,11 @@ from fsspec.registry import register_implementation
 
 from xcube.core.store import (
     DataStoreError,
-    PreloadedDataStore,
+    get_data_store_class,
     list_data_store_ids,
     new_data_store,
 )
+from xcube.core.store.fs.store import BaseFsDataStore
 from xcube.core.store.preload import NullPreloadHandle
 
 
@@ -51,10 +53,20 @@ class ListDataStoreTest(unittest.TestCase):
 
 
 class TestBaseFsDataStore(unittest.TestCase):
+    def test_get_data_types(self):
+        self.assertEqual(
+            {"dataset", "geodataframe", "mldataset"},
+            set(BaseFsDataStore.get_data_types()),
+        )
+
     def test_get_data_opener_ids(self):
         store = new_data_store("file")
         self.assertEqual(
-            ("dataset:geotiff:file",), store.get_data_opener_ids(data_id="test.geotiff")
+            (
+                "mldataset:geotiff:file",
+                "dataset:geotiff:file",
+            ),
+            store.get_data_opener_ids(data_id="test.geotiff"),
         )
         self.assertEqual(
             ("mldataset:geotiff:file",),
@@ -95,6 +107,82 @@ class TestBaseFsDataStore(unittest.TestCase):
         store_test = store.preload_data()
         self.assertTrue(hasattr(store_test, "preload_handle"))
         self.assertIsInstance(store_test.preload_handle, NullPreloadHandle)
+
+
+class FsDataStoreTest(unittest.TestCase):
+    def test_get_filename_extensions_abfs_openers(self):
+        self.assert_accessors("abfs", "openers")
+
+    def test_get_filename_extensions_abfs_writers(self):
+        self.assert_accessors("abfs", "writers")
+
+    def test_get_filename_extensions_file_openers(self):
+        self.assert_accessors("file", "openers")
+
+    def test_get_filename_extensions_file_writers(self):
+        self.assert_accessors("file", "writers")
+
+    def test_get_filename_extensions_ftp_openers(self):
+        self.assert_accessors("ftp", "openers")
+
+    def test_get_filename_extensions_ftp_writers(self):
+        self.assert_accessors("ftp", "writers")
+
+    def test_get_filename_extensions_https_openers(self):
+        self.assert_accessors("https", "openers")
+
+    def test_get_filename_extensions_https_writers(self):
+        self.assert_accessors("https", "writers")
+
+    def test_get_filename_extensions_memory_openers(self):
+        self.assert_accessors("memory", "openers")
+
+    def test_get_filename_extensions_memory_writers(self):
+        self.assert_accessors("memory", "writers")
+
+    def test_get_filename_extensions_s3_openers(self):
+        self.assert_accessors("s3", "openers")
+
+    def test_get_filename_extensions_s3_writers(self):
+        self.assert_accessors("s3", "writers")
+
+    def test_get_filename_extensions_unknown_accessor_type(self):
+        with self.assertRaises(DataStoreError) as dse:
+            self.assert_accessors("s3", "modifiers")
+        self.assertEqual(
+            "Invalid accessor type. Must be 'openers' or 'writers', was 'modifiers'",
+            f"{dse.exception}",
+        )
+
+    def assert_accessors(
+        self, protocol: str, accessor_type: Literal["openers", "writers"]
+    ):
+        store_cls = get_data_store_class(protocol)
+        accessors = store_cls.get_filename_extensions(accessor_type)
+        expected_accessors = {
+            ".geojson": [f"geodataframe:geojson:{protocol}"],
+            ".levels": [f"mldataset:levels:{protocol}", f"dataset:levels:{protocol}"],
+            ".nc": [f"dataset:netcdf:{protocol}"],
+            ".shp": [f"geodataframe:shapefile:{protocol}"],
+            ".zarr": [f"dataset:zarr:{protocol}"],
+        }
+        if accessor_type == "openers":
+            geotiff_openers = {
+                ".geotiff": [
+                    f"mldataset:geotiff:{protocol}",
+                    f"dataset:geotiff:{protocol}",
+                ],
+                ".tif": [
+                    f"mldataset:geotiff:{protocol}",
+                    f"dataset:geotiff:{protocol}",
+                ],
+                ".tiff": [
+                    f"mldataset:geotiff:{protocol}",
+                    f"dataset:geotiff:{protocol}",
+                ],
+            }
+            expected_accessors.update(geotiff_openers)
+        self.assertEqual(accessors, expected_accessors)
 
 
 def test_fsspec_instantiation_error():