+++ xcube/core/store/fs/accessor.py
@@ -4,7 +4,7 @@
 
 import copy
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Optional
 
 import fsspec